task_6_isolate_rag_to_flask_setup/requirements.txt
Flask==3.1.0
python-dotenv==0.19.2
PyJWT==2.3.0
openai
google-generativeai==0.8.3
qdrant-client==1.12.0
pandas==2.2.3
PyPDF2==3.0.1
tiktoken==0.8.0
